requirements.txt
aiohttp==3.6.2
attrs==19.3.0
yarl==1.3.0
